Simplify configuration
Provide an "SSDS" class for staging and release
