+++ tests/test_ssds.py
@@ -7,15 +7,12 @@ import tempfile
 from uuid import uuid4
 from random import randint
 
-import boto3
 from google.cloud import storage
 
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
 import ssds
-from ssds import checksum, aws, s3, gs
-from ssds.config import Staging, Platform
 from tests import infra
 
 
@@ -48,11 +45,11 @@ class TestSSDS(unittest.TestCase, infra.SuppressWarningsMixin):
             with open(os.path.join(root, "large.dat"), "wb") as fh:
                 fh.write(os.urandom(1024 ** 2 * 80))
             with self.subTest("AWS"):
-                with Staging.override(Platform.aws, s3, _s3_staging_bucket):
-                    ssds.upload(root, f"{uuid4()}", "this_is_a_test_submission")
+                with ssds.Staging.override(ssds.Platform.aws, ssds.s3, _s3_staging_bucket):
+                    ssds.Staging.upload(root, f"{uuid4()}", "this_is_a_test_submission")
             with self.subTest("GCP"):
-                with Staging.override(Platform.gcp, gs, _gs_staging_bucket):
-                    ssds.upload(root, f"{uuid4()}", "this_is_a_test_submission")
+                with ssds.Staging.override(ssds.Platform.gcp, ssds.gs, _gs_staging_bucket):
+                    ssds.Staging.upload(root, f"{uuid4()}", "this_is_a_test_submission")
 
 
 class TestSSDSChecksum(infra.SuppressWarningsMixin, unittest.TestCase):
@@ -60,11 +57,11 @@ class TestSSDSChecksum(infra.SuppressWarningsMixin, unittest.TestCase):
         data = b"\x89\xc0\xc6\xcd\xa9$=\xfa\x91\x86\xedi\xec\x18\xcc\xad\xd1\xe1\x82\x8f^\xd2\xdd$\x1fE\x821"
         expected_crc32c = "25c7a879"
         with self.subTest("all at once"):
-            cs = checksum.crc32c(data)
+            cs = ssds.checksum.crc32c(data)
             self.assertEqual(expected_crc32c, cs.hexdigest())
         with self.subTest("sliced"):
             i = randint(1, len(data) - 2)
-            cs = checksum.crc32c(data[:i])
+            cs = ssds.checksum.crc32c(data[:i])
             cs.update(data[i:])
             self.assertEqual(expected_crc32c, cs.hexdigest())
 
@@ -74,15 +71,15 @@ class TestSSDSChecksum(infra.SuppressWarningsMixin, unittest.TestCase):
         with io.BytesIO(data) as fh:
             blob.upload_from_file(fh)
         blob.reload()
-        cs = checksum.crc32c(data).google_storage_crc32c()
+        cs = ssds.checksum.crc32c(data).google_storage_crc32c()
         self.assertEqual(blob.crc32c, cs)
 
     def test_blob_md5(self):
         data = os.urandom(200)
-        blob = aws.resource("s3").Bucket(_s3_staging_bucket).Object("test")
+        blob = ssds.aws.resource("s3").Bucket(_s3_staging_bucket).Object("test")
         with io.BytesIO(data) as fh:
             blob.upload_fileobj(fh)
-        cs = checksum.md5(data).hexdigest()
+        cs = ssds.checksum.md5(data).hexdigest()
         self.assertEqual(blob.e_tag.replace('"', ''), cs)
 
 class TestS3Multipart(infra.SuppressWarningsMixin, unittest.TestCase):
